moving email methods in EmailServiceImpl
