+++ cs-office-365/src/main/java/io/cloudslang/content/office365/services/UserServiceImpl.java
@@ -175,53 +175,4 @@ public class UserServiceImpl {
 
         return finalUrl;
     }
-
-    @NotNull
-    public static Map<String, String> ListAttachment(@NotNull final ListAttachmentsInputs getListAttachmentInputs) throws Exception {
-        final HttpClientInputs httpClientInputs = new HttpClientInputs();
-        final Office365CommonInputs commonInputs = getListAttachmentInputs.getCommonInputs();
-        httpClientInputs.setUrl(ListAttachmentsUrl(commonInputs.getUserPrincipalName(), commonInputs.getUserId(), getListAttachmentInputs.getMessageId()));
-
-        HttpCommons.setCommonHttpInputs(httpClientInputs, commonInputs);
-
-        httpClientInputs.setAuthType(ANONYMOUS);
-        httpClientInputs.setMethod(GET);
-        httpClientInputs.setKeystore(DEFAULT_JAVA_KEYSTORE);
-        httpClientInputs.setKeystorePassword(CHANGEIT);
-        httpClientInputs.setContentType(APPLICATION_JSON);
-
-        httpClientInputs.setResponseCharacterSet(commonInputs.getResponseCharacterSet());
-        httpClientInputs.setHeaders(getAuthHeaders(commonInputs.getAuthToken()));
-
-        return new HttpClientService().execute(httpClientInputs);
-    }
-
-    @NotNull
-    private static String ListAttachmentsUrl(String userPrincipalName, String userId, String messageId) throws Exception {
-        String finalUrl;
-        if (!StringUtils.isEmpty(userPrincipalName))
-            finalUrl = GET_USER_REQUEST_URL + userPrincipalName;
-        else
-            finalUrl = GET_USER_REQUEST_URL + userId;
-        if (!StringUtils.isEmpty(messageId))
-            finalUrl = finalUrl + MESSAGES_PATH + messageId;
-        finalUrl = finalUrl + ATTACHMENTS;
-        return finalUrl;
-    }
-
-    public static String retrieveAttachmentIdList(String returnMessage) {
-        final List<JsonElement> attachmentIdList = new ArrayList();
-        final JsonParser parser = new JsonParser();
-        final JsonObject responseJson = parser.parse(returnMessage).getAsJsonObject();
-        if (responseJson.has(VALUE)) {
-            final JsonArray valueList = responseJson.getAsJsonArray(VALUE);
-            for (int i = 0; i < valueList.size(); i++) {
-                final JsonObject valueObject = (JsonObject) valueList.get(i);
-                if (valueObject.has(ID)) {
-                    attachmentIdList.add(valueObject.get(ID));
-                }
-            }
-        }
-        return StringUtils.join(attachmentIdList, ",");
-    }
 }
